+++ src/components/Card/Card.svelte
@@ -1,11 +1,12 @@
 <script>
   export let hover = true;
-  export let c = "";
+  let className = "";
+  export {className as class};
   export let wrapperClasses = "rounded inline-flex flex-col overflow-hidden";
 </script>
 
 <div
-  class="{c} {hover ? 'elevation-1 hover:elevation-8' : ''} {wrapperClasses}">
+  class="{className} {hover ? 'elevation-1 hover:elevation-8' : ''} {wrapperClasses}">
   <slot name="title" />
   <slot name="media" />
   <slot />