+++ src/components/Checkbox/Checkbox.svelte
@@ -2,7 +2,8 @@
   import Icon from "../Icon";
   import { Ripple } from "../Util";
 
-  export let c = "";
+  let className = "";
+  export {className as class};
   export let value = false;
   export let label = "";
   export let color = "primary";
@@ -16,14 +17,14 @@
   }
 </script>
 
-<div class="{c} {wrapperClasses}" on:click={check}>
+<div class="{className} {wrapperClasses}" on:click={check}>
   <input bind:checked={value} class="hidden" type="checkbox" on:change />
   <div class="relative w-auto h-auto z-0">
     <Ripple color={value && !disabled ? color : 'gray'}>
       {#if value}
-        <Icon c={disabled ? 'text-gray-500' : `text-${color}-500`}>check_box</Icon>
+        <Icon class={disabled ? 'text-gray-500' : `text-${color}-500`}>check_box</Icon>
       {:else}
-        <Icon c={disabled ? 'text-gray-500' : 'text-gray-600'}>
+        <Icon class={disabled ? 'text-gray-500' : 'text-gray-600'}>
           check_box_outline_blank
         </Icon>
       {/if}