+++ src/components/List/ListItem.svelte
@@ -57,7 +57,7 @@
   on:click>
   {#if icon}
     <Icon
-      c="pr-6"
+      class="pr-6"
       small={dense}
       color={selected && navigation ? 'text-primary-500' : ''}>
       {icon}